Merge in the uninstall tester library
Also switch to using PSR-0 autoloading, with a shortened tenor prefix.

In the process we’ve made a lot of the loader methods public that were
protected before, so that the uninstall test case could access them.

In combination with this, it was also necessary to introduce a static
method to retrieve a single instance of the class to be re-used both in
the bootstrap and by the uninstall test case. We’ve also made the
constructor protected as a part of that.

In the future, we’ll likely improve on this state of things
significantly, in regard to passing the instance of the loader around.
It would also be cool if we could automatically detect the plugin to
run the usage simulator and uninstallation for from the loader
directly, instead of still having to define that in the uninstall test
case.

Fixes #1