Run uninstall tests via separate config
This is better than putting them in an excluded group. When they are
placed in an excluded group, they are still considered a part of the
test suite as a while, and so the entire test suite is loaded when they
are run. This isn’t always an issue, but it does cause problems when
some of the test cases in the wide test suite use data providers that
call the plugin’s functions. Since the plugin won’t be loaded during
the uninstall tests, those functions won’t be available then. But the
data providers will still be called, so that PHPUnit can calculate the
total number of tests in the suite. To avoid this, it is best to place
the tests in their own suite. This can be done in a single config file,
but it still requires using an exclude group or else always specifying
the test suite when running the general tests, since you can’t mark one
suite as being the default one that is run when you call `phpunit`. So
using a single config would force us to always run `phpunit —testsuite
default` when running our main suite, and `phpunit —testsuite uninstall
—group uninstall` when running the uninstall tests. It is much easier
to just place them in a separate config, so that we can run the main
tests with `phpunit`, and the uninstall tests with `phpunit -c
phpunit.uninstall.xml.dist`.

See #1